+++ src/main/java/com/agh/surveys/repository/PollRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PollRepository extends JpaRepository<Poll, Long> {
+public interface PollRepository extends JpaRepository<Poll, Integer> {
 
 }
