+++ src/main/java/com/agh/surveys/service/question/IQuestionService.java
@@ -7,13 +7,13 @@ import java.util.List;
 
 public interface IQuestionService {
 
-    List<Question> getByPollId(Long poolId);
+    List<Question> getByPollId(Integer poolId);
 
-    Question addQuestion(Long poolId, QuestionDetails questionDetails);
+    Question addQuestion(Integer poolId, QuestionDetails questionDetails);
 
-    Question getQuestion(Long questionId);
+    Question getQuestion(Integer questionId);
 
-    void deleteQuestion(Long questionId);
+    void deleteQuestion(Integer questionId);
 
     List<Question> addAllQuestionDetails(List<QuestionDetails> questionDetails);
 }
